Crank-Nicolson Method
These scripts solve a system of partial differential equations (an initial boundary-value problem) using the Crank-Nicolson Method. With example initial parameters of the problems defined these scripts solve the problem given a time-frame, produce a 3-dimensional graph of the results and a 2-dimensional plot at 3 key stages of the evolution of the system.