Fisher Equation
A series of scripts which solve the partial differential problem the Fisher equation - using the MATALB PDE solver pdepe. These scripts solve the problem using example boundary and initial conditions and allows to represent the results in a 3-dimensional graph of the evolution of the solution over time, and a 2-dimensional plot of the solution at set time points.